Wrote simple integration tests for the shader.
Needed to make a function in LibraryWrapper for initializing GLFW and GLAD and needed to call it from main.cpp and the integration tests main_test.cpp.  This is messy and needs cleaned up into a unit testable Window class.  I do need to work out how to communicate the dependency of LibraryWrapper on having a current context though.