Add glViewport to IGlfwWrapper and use it from framebuffer size callback.
Currently we use a shortcut for dispatching the framebuffer size callback
because we only allow a single instance of the Window class anyway, but there
are comments about how to dispatch for multiple windows in the future.  Some
details of testability would need to be worked out (likely the dispatch function
would just not be unit testable).

I also renamed the awkward IGlfwWrapper::LoadGLLoader function to just LoadGl.