Make UniqueHandle take pointer to IOpenGLWrapper to be used for deletion.
Previously, the deleter lambdas were capturing the `this` pointer.  This worked
fine until UniqueHandle and the classes using it were made movable.  After this
The target of a move would receive a copy of the deleter from the source of the
move.  However, since that deleter referenced the `this` pointer of the source
rather than just the IOpenGLWrapper it would segfault even if the IOpenGLWrapper
was still valid.  This is because it was capturing the `this` pointer and then
looking up the IOpenGLWrapper property in the object pointed to by the captured
`this` rather than having a pointer to the IOpenGLWrapper directly.

One possible solution to this was to just capture an IOpenGLWrapper pointer in
the lambda.  However, I thought it would be less error prone and more explicit
to have the UniqueHandle directly depend on an IOpenGLWrapper and pass it as an
argument to the deleter.